Do not hardcode config dir
